+++ src/PCEHR.Sample/UploadDocumentClientSample.cs
@@ -49,6 +49,8 @@ namespace PCEHR.Sample
 
             // Create PCEHR header
             CommonPcehrHeader header = PcehrHeaderHelper.CreateHeader();
+            // Override this value to the current patient's IHI.
+            header.IhiNumber = "IHI";
 
             // Create the client
             // SVT endpoint is https://b2b.ehealthvendortest.health.gov.au/uploadDocument
@@ -75,8 +77,8 @@ namespace PCEHR.Sample
                 packageBytes,
                 "formatCode",       
                 "formatCodeName",
-                HealthcareFacilityTypeCodes.Transport,
-                PracticeSettingTypes.SpecialistMedicalPractitionerServiceNEC
+                HealthcareFacilityTypeCodes.GeneralPractice,                // Update to relevant code
+                PracticeSettingTypes.GeneralPracticeMedicalClinicService    // Update to relevant code
                 );
 
             // To supercede / amend an existing document, the same UploadDocument call is used. However, the request is 
@@ -90,8 +92,8 @@ namespace PCEHR.Sample
             //    packageBytes,
             //    "formatCode",
             //    "formatCodeName",
-            //    HealthcareFacilityTypeCodes.Transport,
-            //    PracticeSettingTypes.SpecialistMedicalPractitionerServiceNEC,
+            //    HealthcareFacilityTypeCodes.GeneralPractice,
+            //    PracticeSettingTypes.GeneralPracticeMedicalClinicService,
             //    "uuidOfDocumentToReplace"
             //    );