+++ src/PCEHR/Enums/ClassCodes.cs
@@ -109,7 +109,7 @@ namespace Nehta.VendorLibrary.PCEHR
         HealthCheckScheduleView,
 
         /// <summary>
-        /// Medicare Overview
+        /// Observation Overview
         /// </summary>
         [CodedValueAttribute(CodingSystem = "NCTIS", CodingSystemOID = "1.2.36.1.2001.1001.101", ConceptCode = "100.16872",
             ConceptName = "Observation View", AlternateName = "Observation View")]
@@ -189,8 +189,25 @@ namespace Nehta.VendorLibrary.PCEHR
         /// Advance care information class code.
         /// </summary>
         [CodedValueAttribute(CodingSystem = "NCTIS Data Components", CodingSystemOID = "1.2.36.1.2001.1001.101", ConceptCode = "100.16975",
-            ConceptName = "Advance Care Information", AlternateName = "Advance Care Planning Document")]
+            ConceptName = "Advance Care Information", AlternateName = "Advance Care Information")]
         AdvanceCareInformation,
+
+        /// <summary>
+        /// AdvanceCareInformation Subtype
+        /// Advance care planning type code (NOT CLASS CODE).
+        /// </summary>
+        [CodedValueAttribute(CodingSystem = "NCTIS Data Components", CodingSystemOID = "1.2.36.1.2001.1001.101", ConceptCode = "100.16998",
+            ConceptName = "Advance Care Planning Document", AlternateName = "Advance Care Planning Document")]
+        AciTypeAdvanceCarePlanning,
+
+        /// <summary>
+        /// AdvanceCareInformation Subtype
+        /// Goals Of Care Document type code (NOT CLASS CODE).
+        /// </summary>
+        [CodedValueAttribute(CodingSystem = "NCTIS Data Components", CodingSystemOID = "1.2.36.1.2001.1001.101", ConceptCode = "100.32016",
+            ConceptName = "Goals of Care Document", AlternateName = "Goals of Care Document")]
+        AciTypeGoalsOfCare,
+
 
         // CeHR documents
 
@@ -230,12 +247,32 @@ namespace Nehta.VendorLibrary.PCEHR
         BirthDetails,
 
         /// <summary>
-        /// Medicare Overview
+        /// Medicines View
         /// </summary>
         [CodedValueAttribute(CodingSystem = "NCTIS", CodingSystemOID = "1.2.36.1.2001.1001.101", ConceptCode = "100.32002",
             ConceptName = "Medicines View", AlternateName = "Medicines View")]
         MedicinesView,
 
+        /// <summary>
+        /// Pharmacist Shared Medicines List
+        /// </summary>
+        [CodedValueAttribute(CodingSystem = "LOINC", CodingSystemOID = "2.16.840.1.113883.6.1", ConceptCode = "56445-0",
+            ConceptName = "Pharmacist Shared Medicines List", AlternateName = "Pharmacist Shared Medicines List")]
+        PharmacistSharedMedicinesList,
+
+        /// <summary>
+        /// Pathology Overview
+        /// </summary>
+        [CodedValueAttribute(CodingSystem = "NCTIS", CodingSystemOID = "1.2.36.1.2001.1001.101", ConceptCode = "100.32026",
+            ConceptName = "Pathology Overview", AlternateName = "Pathology Overview")]
+        PathologyOverview,
+
+        /// <summary>
+        /// Diagnostic Imaging Overview
+        /// </summary>
+        [CodedValueAttribute(CodingSystem = "NCTIS", CodingSystemOID = "1.2.36.1.2001.1001.101", ConceptCode = "100.32025",
+            ConceptName = "Diagnostic Imaging Overview", AlternateName = "Diagnostic Imaging Overview")]
+        DiagnosticImagingOverview,
 
         /*
         /// <summary>